pyproject.toml
[project]
name = "stk-mcp"
version = "0.2.0"
description = "STK-MCP, an MCP server allowing LLMs to interact with Ansys/AGI STK - Digital Mission Engineering Software"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.7",
    "typer>=0.15.2",
    "uvicorn>=0.30",
    "rich>=13.7",
    "pywin32>=306; platform_system == 'Windows'",
    "agi-stk12",
    "pydantic-settings>=2.10.1",
    "tenacity>=9.1.2",
]

[project.scripts]
stk-mcp = "stk_mcp.cli:app" # New CLI entry point

# Optional: Add pywin32 if needed for clarity, though pip usually handles it on Windows
# dependencies = [
#     "mcp[cli]>=1.6.0",
#     "pywin32>=306; sys_platform == 'win32'" # Example conditional dependency
# ]

[build-system]
requires = ["uv-build>=0.4.0"]
build-backend = "uv_build"

[tool.uv.sources]
agi-stk12 = { path = "agi.stk12-12.10.0-py3-none-any.whl" }
